fix: lint

+++ modules/frontend/auth/data/src/services/interceptors.ts
@@ -12,8 +12,7 @@ export function configAxiosInstance(instance: AxiosInstance) {
     (error) => {
       console.warn({ error });
 
-      if (error.response?.status === 401)
-        authCookiesService.deleteCookieClient();
+      if (error.response?.status === 401) authCookiesService.deleteCookieClient();
 
       return Promise.reject(error);
     }
@@ -24,10 +23,7 @@ export function configAxiosInstance(instance: AxiosInstance) {
 export function Log(context: string) {
   return {
     apiError(functionName: string, error: Error) {
-      console.warn(
-        `[API] ${context}.${functionName} : ${error.name} - ${error.message} \n\n${error.stack}\n\n`,
-        error
-      );
+      console.warn(`[API] ${context}.${functionName} : ${error.name} - ${error.message} \n\n${error.stack}\n\n`, error);
     },
     unexpectedError(functionName: string, error: Error) {
       const message = `[JS] AuthService.${functionName} : ${error.name} - ${error.message} \n\n${error.stack}`;
